PY310-OCT-VL/requirements.txt
numpy==1.26.4
scipy==1.11.3
PyQt5==5.15.10
pyqtgraph==0.13.3
seabreeze==2.6.0
pyserial==3.5
typing_extensions==4.7.1
